requirements.txt
# fundamental packages
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
python-dotenv
tabulate

# crawling
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.1

# schedule
schedule==1.2.1
apscheduler==3.10.4

# pyspark
pyspark==3.5.0
findspark==2.0.1

# machine learning
scikit-learn==1.3.2
implicit==0.7.2

# notebook
jupyterlab==4.1.5
ipykernel==6.29.0

# visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0


